Ensure that jinja loader is not registered at each render
The env.loader was replaced by a new ChoiceLoader wrapping the previous
loader on every render. It seems that the `_loaded` attribute was here
to avoid re-registering the loader but it was set to `True` on the
handler instance and not on its class making it useless since the a
handler instance is created for each request.

Closes #114.